requirements.txt
streamlit==1.28.1
Pillow==10.0.1
numpy==1.24.3
pandas==1.5.3
plotly==5.17.0
requests==2.31.0
pandoc==2.19.2